Remove --enable-plugins=list variation.
With the removal of the 'experimental' plugin group, passing
--enable-plugins=experimental to configure will now cause the Makefile
to try and build a plugin named 'experimental', which will result in a
mysterious "experimental: No such file or directory" build error.

This is because of the "list" fallback behavior: if the option is not
one of "no", "base" or "all" then the configure script assumes it to be
a list of plugin names.  I've never known that to be useful to anyone,
so remove the "list of plugin names" behavior and replace it with an
informative error message.

Eventually the --enable-plugins will be removed entirely since EPlugin
is deprecated.