Display local contact photos in EContactEditor's EImageChooser
Some contacts have photos stored locally as files. Until now
EContactEditor displayed thumbnail of contact photo in EImageChooser
only if the photo was embedded in the vCard as data. This patch
adds support for dispalying thumbnails of photos stored in local files
to indicate that the contact has a picture set.