Delay saving Evolution version until later in initialization.
Save the version after the startup wizard has had a chance to
run.  If the user chooses to restore data and settings from a
backup, Evolution will restart and the restored data may need
to be migrated.

If we save the version before the restart, then Evolution will
think it has already migrated data and settings to the current
version and the restored data may not be handled properly.