Remove mail_ui_session_source_changed_cb().
This function removed a CamelStore from EMFolderTreeModel and then added
it again in response to the corresponding ESource emitting a "changed"
signal (presumably from editing a mail account and saving changes).

In the past this was the only way for changes to the account URI string
to take effect without restarting Evolution.  But we have CamelSettings
now and don't need this hack anymore.  Time for it to die.