Disconnect a CamelService after editing its settings.
This will force Evolution to reconnect to the service using the current
settings.  However this is not a complete solution.  If the new settings
now point to a completely different mail account, we leave behind cached
messages and database tables from the previous account such that you end
up with some weird hybrid of the previous account and current account.

I guess for now the answer is "don't do that", but we should try to
handle that more gracefully in the future -- more for architectural
correctness than it being a common real world use case.