Remove all GDK threads usage.
According to [1], we don't need to worry about GDK's global lock since
we don't call gdk_threads_init() or gdk_threads_set_lock_functions().

The GDK threads API is being aggressively deprecated by GTK+ developers
so let's just abandon it entirely.  I've never really understood when
you're supposed to use it or not use it anyway, so it's good to be rid
of this confusion.

[1] https://mail.gnome.org/archives/desktop-devel-list/2012-August/msg00005.html