Return of the "mark messages as read" preference.
This reverses the removal of the "Mark messages as read" preference in
commit e7247d6d60336861ed5c0ba05dd64e29f728349c.

I apparently misjudged the popularity of this preference, as there's
been a surprising amout of backlash and confusion as to its removal on
the mailing list, IRC channel, and various distro-support forums.

So, "mea culpa" for removing that one.  Haven't heard a peep about the
other removed preferences, however, so they will stay gone.