backup-restore: Add version info to evolution.dirs file.
Store the Evolution version used to create the backup file in the
evolution.dir file which is part of the tarball.

After restoring the backup file we need to reset the last-used-version
key in GSettings to the version listed in the evolution.dir file so the
proper migration routines run when Evolution is restarted.