Run mbox-to-Maildir conversion before loading modules.
Invoke the mbox-to-Maildir conversion directly from main(), just
before the call to e_shell_load_modules().

The reason the code is here and not in the mail module is because
we inform the user at startup of the impending mail conversion by
displaying a popup dialog and waiting for confirmation.

This has to be done before we load modules because some of the
EShellBackends immediately add GMainContext sources that would
otherwise get dispatched during gtk_dialog_run(), and we don't
want them dispatched until after the conversion is complete.