Remove the settings capplet.
The evolution-settings capplet was originally designed for Anjal, it was
used in MeeGo as part of the Express Mode effort, but doesn't really fit
in GNOME 3 nowadays (nor did it really fit in GNOME 2, in my opinion).

This is pretty clearly dead weight at this point.  The MeeGo developers
have disappeared, and the remaining Evolution developers are not and do
not intend to maintain it.  Plus it doesn't even build currently.