Check for SpamAssassin during configure.
If the spamassassin and sa-learn programs cannot be found with
AC_PATH_PROG, configure will abort with an error message.

You can then either:
a) install the SpamAssassin software
b) specify the path with SPAMASSASSIN=/path/to/spamassassin and/or
   SA_LEARN=/path/to/sa-learn
c) pass --disable-spamassassin to configure to exclude the module

This also drops the "spamc" and "spamd" GSettings keys.