Remove e_mail_junk_filter_available().
Because we now check for junk filtering software in configure, junk
filtering modules can now safely assume the required junk filtering
software is installed.  No more having the module installed but not
the required software.  If this invariant is broken, an error will
be shown for each new message received.