em_folder_tree_model_get_folder_name: Don't return NULL.
If a folder's display name is not available in the folder tree model for
some reason, return the raw folder path rather than NULL.  This silences
a runtime warning in the mail-notification plugin.

This feels like a Band-Aid for a deeper issue -- perhaps we're racing
the folder tree model?  Might be better to just fetch the CamelFolder
asynchronously and get the folder's display name directly.