Calendar's Work Week view could left empty after start
When Evolution started in the Calendar view with Work Week view
selected it left the view empty, because the Work Week view was
not set focused. This change was not made, because GnomeCalendar
has set Work Week view as a default view, thus gnome_calendar_set_view()
just returned for "no change made" on the view type. Checking also
'in_focus' property of the view makes sure the initialization will
be done as expected.