[prefer-plain] Always show HTML part as attachment for only HTML messages
When the prefer-plain is set to show only text/plain parts, without showing
suppressed HTML parts as attachments, and the received message contains
only one part, the text/html part, then the preview was completely empty,
which is confusing. With this change, the text/html only messages will
show the suppressed part as attachment regardless the setting.