Bug #684447 - Check for highlight during configure.
If the highlight program cannot be found width AC_PATH_PROGS,
configure will abort with an error message.

You can either
a) install highlight
b) specify the patch with HIGHLIGHT=/path/to/highlight
c) pass --disable-text-highlight to configure to exclude the module

This also makes text-highlight module to fallback to text/plain
formatter when highlight program would crash or fail to ensure
the content is delivered to user.