Bug 684691 - EMailConfigFormatHTML: Signal handler never disconnected
EMailConfigFormatHTML listens for "changed" signals from a GSettings
instance, passing itself as the 'user_data' to the signal handler.  But
the signal handler was left connected after EMailConfigFormatHTML was
finalized, resulting in the signal handler receiving a dangling pointer.

Not using g_signal_connect_object() here because of the unresolved
reference leak issue in GObject.  The GSettings instance is likely
cached internally and lives well beyond the EMailConfigFormatHTML.