EMailAutoconfig: Restructure XML parsing.
If we get a hit from the online database, parse the XML immediately and
store the results in generic internal structs.  Then when we're given an
ESource to populate we can feed it the stored results instead of parsing
the XML on demand (and repeatedly).

What this is really doing is making room for other auto-configuration
methods besides the online database lookup.