Bug 687137 - Google account user name overwritten with email address
Evolution's Google account module was setting the ECollectionSource's
"identity" property to the account's email address before submitting the
account to E-D-S.  Meanwhile, the Google module in E-D-S was syncing the
account user name to the ECollectionSource's "identity" property.

This mismatch resulted in an explicitly specified GMail user name being
overwritten when the account information is submitted.  Evolution should
instead set the "identity" property to the account's user name.