Reset the maintainer list in the About dialog.
No disrespect to past maintainers, documenters and contributers, but the
Credits section of Evolution's About dialog is a mess and needs a reset.
The list of authors and documenters is YEARS out of date, and it's not
feasible to maintain a complete list of contributors for a project this
large and this old.

Reset the authors list to myself, Milan and Dan and the documenters list
to Andre -- with a nod to past contributers.