Add a priority field for mail formatter and parser extensions.
Add a priority field to EMailFormatterExtension and EMailParserExtension
class structs.  Extension classes can then explicitly specify a priority
with respect to other extension classes with the same MIME type, so that
the order of extension registration doesn't matter.

Priority field defaults to G_PRIORITY_DEFAULT.  Built-in formatters and
extensions will use G_PRIORITY_LOW.  We can get more sophisticated with
priority values if we need to, but this should suffice for now.