Remove --express command-line option.
This effectively removes Express Mode as a user-selectable feature.

Express Mode was developed for MeeGo by MeeGo developers.  Today the
MeeGo project is dead, the MeeGo developers have disappeared, no one
is maintaining Express Mode and the feature is deteriorating.

There's still some bits of Express Mode I'd like to salvage before
we remove it entirely, but henceforth it will have to be enabled at
compile-time.