e_mail_folder_uri_parse: Handle local mbox folder URIs.
In Evolution 2.x, the local mail store used mbox format.  Camel will not
recognize the old-style "mbox:///.../local" folder URIs, since the local
mail store is now Maildir format.  Test for this and work around it.

This should re-fix bug 638307 which was a nuisance error about failing
to append to a "mbox:///.../local#Sent" folder after sending a message.