alarm-queue: Remove the "mail notifications not supported" dialog
This dialog is all kinds of lame, not to mention ugly.  It tends to pop
up when evolution-alarm-notify starts and often steals input focus from
the normal reminder dialog that pops up overtop of it.

Let's not apologize for features not implemented.  Instead of alerting
the user that we'll show a normal reminder dialog in lieu of sending an
email reminder, just show a normal reminder dialog.