Remove "Search for sender photograph only in local address books".
This was added as part of bug 360184 but no justification was given
for the "local-only" part.  My Spidey sense tells me it was a hack-
around for the old implementation's tendency to freeze the UI while
searching for a photograph.  So the "local-only" option really just
meant "don't freeze the UI for very long, please".

The new EPhotoCache-based implementation in 3.8 NEVER freezes the UI,
so the "local-only" option is no longer needed.  If a remote address
book is slow or unresponsive we simply cancel the async photo lookup
when the user moves on to another email.