Add contact-photos module.
This encapsulates the EContactPhoto look up feature that was previously
built into EPhotoCache.  It's now implemented as an EPhotoSource -- one
per address book.  One advantage of this implementation is that address
books are now queried concurrently rather than serially.

EPhotoCacheContactLoader is an EPhotoCache extension that takes care of
adding and removing EPhotoSources for available address books.