Make sure EAddressbookModel has a fresh EBookClient.
In the event of an address book backend abort, EClientCache detects this
and invalidates its cached EClient (if it has one), so a new instance is
created on the next request.

EAddressbookModel is only handed an EClient once, which may become stale
if the backend aborts.  And even if the backend is restarted the address
book will remain unresponsive in Evolution.

This commit changes the behavior so that every time an address book is
selected in the side bar, a fresh EClient instance is obtained from the
EClientCache and handed to the EAddressbookModel.  If the model already
has that EClient instance, nothing happens.  Otherwise the model resets
itself and creates a new EBookClientView.

(cherry picked from commit 6ec48e33ef01476246d943d941eb2e022687ef59)