Bug 697632 - [mail-to-task] Invalid unref of a CamelFolder
e_mail_reader_get_folder() does not return a new CamelFolder reference,
yet mail_to_event() was acting as though it does.  This caused a crash
after the function ran and Evolution tried to use the folder again.

Truth be told, e_mail_reader_get_folder() really *should* return a new
reference to ensure the CamelFolder is not finalized while it's in use.
But we would need to rename the function to e_mail_reader_ref_folder()
to reflect the change in semantics, and I suspect the function is used
in a great many places.