Distinguish "attachment-wrapper" elements from attachments.
The HTML for attachments always has the following form:

<div class="attachment-wrapper" id="something" style="display: block;">
  <actual attachment element>
</div>

The <div> element controls attachment visibility through its "display"
style attribute, which is either "block" or "none".

Problem is the <actual attachment element> was getting the same ID as
its parent <div> element.  So when either element was requested by ID,
in certain cases the wrong element was returned and caused misbehavior
and console warnings.

Solve this by adding a "wrapper" suffix to the <div> element ID.  So in
the example above, id="something" gets the <actual attachment element>,
whereas id="something.wrapper" gets the <div> element.