EMailReader: Add a "composer-created" signal.
Emitted to indicate a composer window was created in response to a user
action on the EMailReader.  Examples of such actions include replying,
forwarding, and starting a new message.

If applicable, the source message (i.e. the message being replied to or
forwarded) is included in the signal emission.

New functions:

  e_mail_reader_composer_created()