Bug #689806 - [webkit-composer] Save as draft forgets signature position
When opening message that has signature with 'Edit as new message', no signature
is added because in signature combo box is active signature set to None. We need
to set active signature id in combo box to that signature that message contains.

Also insert signatures through DOM manipulation, because we don't want to insert
empty lines that WebKit inserts through 'insertParagraph' command.