Bug 702454 - EMailBrowser fails to show message
This reworks the MessageList regen scheduling to be a little more
intuitive, at least to me.  We now set the RegenData immediately, but
start the actual regen operation from an idle callback.  That way the
caller has the remainder of this main loop iteration to make further
MessageList changes without triggering additional regens.

I think what was happening before was we were triggering multiple regen
operations as we were configuring the EMailBrowser, with each new regen
cancelling the previous, and the message UID to select wound up getting
paired with one of the cancelled regen operations.  This resulted in no
message UID ever getting selected in the EMailBrowser.