User docs: Add useful section link titles for sections of the "account management" guides
This avoids having ten undistinguishable links to with exactly the same names in the 'More Information' section at the bottom of the Yelp help browser for all those Account Management topic pages. Instead they know mention which account type they specifically link to.

Thanks to Kat and Shaun for the help.