Add ETreeViewFrame.
ETreeViewFrame embeds a GtkTreeView in a scrolled window and adds an
inline-style toolbar beneath the scrolled window which can be hidden.

The inline-style toolbar supports "add" and "remove" actions, as well
as move actions if the tree view is reorderable and selection actions
if the tree view supports multiple selections.  The action set can be
extended through e_tree_view_frame_insert_toolbar_action().

This also adds a small demo program: test-tree-view-frame