GalViewCollection: We don't need no stinkin' factories!
Given a type code string from an XML file, find the appropriate GType by
traversing the GType hierarchy from GAL_TYPE_VIEW and checking the class
structures for a matching type code string.

This completely eliminates the need for what's left of GalViewFactory.
Now it's just a matter of cleaning up the remains.