Check if returned EEditor is valid.
When closing down an editor with attachments added the editor is destroyed and
the attachments are freed. But from store the row-deleted signal is emitted for
every attachment. In callback for that signal - attachment_store_changed_cb
we are marking the editor as changed. But the e_msg_composer_get_editor will return
us the NULL value instead of valid address as EEditor is already destroyed.