Add e_web_view_redirect_uri().
Replaces the given URI with a redirected URI as necessary, primarily for
use with custom SoupRequest handlers.  Typically this function would be
called just prior to handing a request off to a SoupSession, such as
from a WebKitWebView "resource-request-starting" signal handler.

Case in point: EMailDisplay now implements the redirect_uri() method,
and calls it from its own "resource-request-starting" signal handler.