Add e_web_view_cursor_image_copy().
Asynchronously copies the image under the cursor to the clipboard.

This replaces the "cursor-image" property, which attempts to match the
image URI to a subresource from WebKitWebDataSource.  The problem with
that approach is EMailDisplay redirects several URI schemes to its own
custom request handlers which further mutate the URI.  So for example,
a text/html message may use a "cid:" URI to refer to an embedded image,
which EMailDisplay transforms to a "mail:" URI and finally to a "data:"
URI.  The final image URI might not be derivable from the original URI
without retracing the SoupRequest.

The "image-copy" action now calls e_web_view_cursor_image_copy().

This also adds an explicit requirement on gdk-pixbuf-2.0 >= 2.24 for
gdk_pixbuf_new_from_stream_async() / finish().

New functions:

  e_web_view_cursor_image_copy()

Removed functions:

  e_web_view_get_cursor_image()
  e_web_view_set_cursor_image()