Add e_web_view_suggest_filename().
Attempts to derive a suggested filename from the given URI for use in a
"Save As" dialog.

By default the suggested filename is the last path segment of the given
URI (the unless the given URI looks like a query), but subclasses can use
other mechanisms for custom URI schemes.  For example, "cid:" URIs in an
email message may refer to a MIME part with a suggested filename in its
Content-Disposition header.