Add e_web_view_new_activity().
Returns a new EActivity for an EWebView-related asynchronous operation,
and emits a "new-activity" signal.  By default the EActivity comes loaded
with a GCancellable and sets the EWebView itself as the alert sink (which
means alerts are displayed directly in the content area).  The signal
emission allows the EActivity to be further customized and/or tracked by
the application.