Bug 709428 - Searchbar widgets should not wrap
Instead of wrapping the searchbar for small screens (which looks ugly
and breaks the initial window size on large screens), hide the filter
combo box in views that would otherwise be too wide for the screen.

There's no loss of functionality when hiding the filter combo box.
It's just a set of convenient pre-defined searches, all of which can
be reproduced through the Advanced Search interface.

New functions:

  e_shell_searchbar_get_filter_visible()
  e_shell_searchbar_set_filter_visible()