Fix error handling glitch in mail_session_send_to_thread().
If sending a message fails but the user was already connected to the
MTA, the error is missed and not handled until further into the logic,
where it's treated as though sending succeeded and only post-processing
failed.  This results in the user seeing a copy of the message in Sent,
but the message was never actually sent.

Fallout from the investigation of bug 710807.