Bump webkitgtk-3.0 requirement to 2.0.1.
Between 2.0.0 and 2.0.1, webkit_dom_html_style_element_get_sheet()
silently changed from returning a WebKitDOMStyleSheet to returning
a WebKitDOMCSSStyleSheet.

EWebView already depends on this newer behavior, which effectively means
Evolution's WebKitGTK+ dependency is already 2.0.1 or later.  Update the
minimum webkitgtk-3.0 requirement in configure.ac to reflect this.

WebKitGTK+ 1.10 is old enough now that the bump should not be an issue.